draft the cleaning
