+++ src/shared/bodies/solid_body.cpp
@@ -7,17 +7,5 @@
 namespace SPH
 {
 //=================================================================================================//
-SolidBodyPartForSimbody::
-    SolidBodyPartForSimbody(SPHBody &body, Shape &body_part_shape)
-    : BodyRegionByParticle(body, body_part_shape),
-      rho0_(DynamicCast<Solid>(this, body.getBaseMaterial()).ReferenceDensity()),
-      Vol_(base_particles_.getVariableDataByName<Real>("VolumetricMeasure")),
-      pos_(base_particles_.getVariableDataByName<Vecd>("Position"))
-{
-    setMassProperties();
-}
-//=================================================================================================//
-SolidBodyPartForSimbody::SolidBodyPartForSimbody(SPHBody &body, SharedPtr<Shape> shape_ptr)
-    : SolidBodyPartForSimbody(body, *shape_ptr.get()) {}
 //=================================================================================================//
 } // namespace SPH
