+++ src/shared/simbody_sphinxsys/body_part_for_simbody.cpp
@@ -1,0 +1,21 @@
+#include "body_part_for_simbody.h"
+
+#include "base_particles.hpp"
+
+namespace SPH
+{
+//=================================================================================================//
+SolidBodyPartForSimbody::
+    SolidBodyPartForSimbody(SPHBody &body, Shape &body_part_shape)
+    : BodyRegionByParticle(body, body_part_shape),
+      rho0_(DynamicCast<Solid>(this, body.getBaseMaterial()).ReferenceDensity()),
+      Vol_(base_particles_.getVariableDataByName<Real>("VolumetricMeasure")),
+      pos_(base_particles_.getVariableDataByName<Vecd>("Position"))
+{
+    setMassProperties();
+}
+//=================================================================================================//
+SolidBodyPartForSimbody::SolidBodyPartForSimbody(SPHBody &body, SharedPtr<Shape> shape_ptr)
+    : SolidBodyPartForSimbody(body, *shape_ptr.get()) {}
+//=================================================================================================//
+} // namespace SPH
