+++ src/shared/simbody_sphinxsys/xml_engine.h
@@ -32,8 +32,8 @@
 #include "SimTKcommon/internal/Xml.h"
 #include "base_data_package.h"
 #include "simbody_middle.h"
+#include "simtk_wrapper.h"
 #include "sphinxsys_containers.h"
-#include "type_wrapper.h"
 
 #include <filesystem>
 #include <fstream>
@@ -51,7 +51,7 @@ class XmlEngine
     /** Constructor for XML output.  */
     XmlEngine(const std::string &xml_name, const std::string &root_tag);
     /** Default destructor. */
-    virtual ~XmlEngine(){};
+    virtual ~XmlEngine() {};
 
     SimTK::Xml::Element root_element_; /**< Root element of document. */
